dlt_airflow/requirements.txt
dlt[bigquery]>=0.5.1
pandas
google-cloud-bigquery
google-cloud-bigquery-storage
pendulum
apache-airflow-providers-google
pyarrow